+++ module_animal.py
@@ -1,15 +1,10 @@
 class Animal:
-    def __init__(self, name, type, specie, weight):
+    def __init__(self, name, type, specie, weight, predator):
         self.name = name
         self.type = type
         self.specie = specie
         self.weight = weight
-    
-    def move(self):
-        pass
-    
-    def eat(self):
-        pass  
+        self.predator = predator
           
     def info(self):
-        print(f'- {self.name} the {self.specie}, is a {self.type} and it weights {self.weight} lb')
+        print(f'- {self.name} is a {self.specie}, is a {self.type} and it weights {self.weight} lb')
